refactor: clean up code warnings and improve code quality
Backend (Rust):
- Remove unused manifest key `workspace.default-run` from Cargo.toml
- Fix unused_must_use warning in server initialization
- Remove unused imports (std::sync::Arc, std::collections::HashMap, etc.)
- Add #[allow(dead_code)] to utility functions kept for future use
- Fix PhantomData import issue in backend_plugin.rs
- Run cargo fix and clippy --fix to auto-fix minor issues

Frontend (TypeScript):
- Remove unused PublicRoute component from App.tsx
- Remove unused Save icon import from SimpleRuleBuilderSplit.tsx
- Fix unused variables (_min, _max, _operator, _threshold, _t)
- Remove debug console.log statements

+++ crates/agent/src/agent/intent_classifier.rs
@@ -685,7 +685,7 @@ impl IntentClassifier {
             score += 0.25;
         }
 
-        (score as f32).min(1.0).max(0.0)
+        score.min(1.0).max(0.0)
     }
 
     /// 告警通道意图评分
@@ -742,7 +742,7 @@ impl IntentClassifier {
             }
         }
 
-        (score as f32).min(1.0).max(0.0)
+        score.min(1.0).max(0.0)
     }
 
     /// 构建澄清结果
